refactor: moved the area search into it's own components
